backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
whisper==1.0.2
openai-whisper==20231117
groq==0.4.1
gtts==2.4.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
librosa==0.10.0
soundfile==0.12.1
